+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ83AQQVJZ4RF1WEWHH0TXBY"
+        "01KJ89283VE2KNXNHB1GZCG84N"
     )
     r"""The unique identifier of the data source"""