+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ83AQR39BSR2Y86WD2NYM6A"
+        "01KJ892843A1RE413WMXN4RBY5"
     )
     r"""The unique identifier of the data source"""