+++ packages/orq-rc/RELEASES.md
@@ -6069,3 +6069,13 @@ Based on:
 - [python v4.5.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.1 - packages/orq-rc
+
+## 2026-02-24 16:55:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.728.0 (2.839.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.2 - packages/orq-rc